IAM-INTEGRATION
Each service will now communicate to IAM, the resource ID and action ID as two additional parameters in token validation API (new Authorize API). IAM will verify based on its policies and respond to the service with all the details that token API did, with an additional “Yes” or “No” response- whether a user is allowed to perform the action on that specific resource.

To cater to these new IAM changes, this middleware is added; it communicates to IAM, the action and the associated resource corresponding to each API.

https://blueprints.launchpad.net/jio/+spec/sbs-iam-integration

Implements bp sbs-iam-integration